Add SV prediction code
