--- python/scoring_model.py
@@ -1,6 +1,0 @@
-"""
-scoring_model.py: Score the structural variants using the binary classification model.
-"""
-
-import os
-import sys
